Added version 3.3.2022

+++ chromium/chrome/browser/extensions/api/browsing_data/browsing_data_api.h
@@ -116,7 +116,7 @@ class BrowsingDataRemoverFunction
   // based on the API call they represent.
   // Returns whether or not removal mask retrieval was successful.
   // |removal_mask| is populated with the result, if successful.
-  virtual bool GetRemovalMask(int* removal_mask) = 0;
+  virtual bool GetRemovalMask(uint64_t* removal_mask) = 0;
 
   // Returns true if the data removal is allowed to pause Sync. Returns true by
   // default. Subclasses can override it to return false and prevent Sync from
@@ -135,7 +135,7 @@ class BrowsingDataRemoverFunction
   // that can be used with the BrowsingDataRemover.
   // Returns true if parsing was successful.
   bool ParseOriginTypeMask(const base::DictionaryValue& options,
-                           int* origin_type_mask);
+                           uint64_t* origin_type_mask);
 
   // Parses the developer-provided list of origins into |result|.
   // Returns whether or not parsing was successful. In case of parse failure,
@@ -152,8 +152,8 @@ class BrowsingDataRemoverFunction
   void OnTaskFinished();
 
   base::Time remove_since_;
-  int removal_mask_ = 0;
-  int origin_type_mask_ = 0;
+  uint64_t removal_mask_ = 0;
+  uint64_t origin_type_mask_ = 0;
   std::vector<url::Origin> origins_;
   content::BrowsingDataFilterBuilder::Mode mode_ =
       content::BrowsingDataFilterBuilder::Mode::BLACKLIST;
@@ -174,7 +174,7 @@ class BrowsingDataRemoveAppcacheFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveAppcacheFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveFunction : public BrowsingDataRemoverFunction {
@@ -185,7 +185,7 @@ class BrowsingDataRemoveFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
   bool IsPauseSyncAllowed() override;
 };
 
@@ -198,7 +198,7 @@ class BrowsingDataRemoveCacheFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveCacheFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveCookiesFunction : public BrowsingDataRemoverFunction {
@@ -210,7 +210,7 @@ class BrowsingDataRemoveCookiesFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveCookiesFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveDownloadsFunction : public BrowsingDataRemoverFunction {
@@ -222,7 +222,7 @@ class BrowsingDataRemoveDownloadsFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveDownloadsFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveFileSystemsFunction
@@ -235,7 +235,7 @@ class BrowsingDataRemoveFileSystemsFunction
   ~BrowsingDataRemoveFileSystemsFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveFormDataFunction : public BrowsingDataRemoverFunction {
@@ -247,7 +247,7 @@ class BrowsingDataRemoveFormDataFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveFormDataFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveHistoryFunction : public BrowsingDataRemoverFunction {
@@ -259,7 +259,7 @@ class BrowsingDataRemoveHistoryFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveHistoryFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveIndexedDBFunction : public BrowsingDataRemoverFunction {
@@ -271,7 +271,7 @@ class BrowsingDataRemoveIndexedDBFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveIndexedDBFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveLocalStorageFunction
@@ -284,7 +284,7 @@ class BrowsingDataRemoveLocalStorageFunction
   ~BrowsingDataRemoveLocalStorageFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemovePluginDataFunction
@@ -297,7 +297,7 @@ class BrowsingDataRemovePluginDataFunction
   ~BrowsingDataRemovePluginDataFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemovePasswordsFunction : public BrowsingDataRemoverFunction {
@@ -309,7 +309,7 @@ class BrowsingDataRemovePasswordsFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemovePasswordsFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveServiceWorkersFunction
@@ -322,7 +322,7 @@ class BrowsingDataRemoveServiceWorkersFunction
   ~BrowsingDataRemoveServiceWorkersFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveCacheStorageFunction
@@ -335,7 +335,7 @@ class BrowsingDataRemoveCacheStorageFunction
   ~BrowsingDataRemoveCacheStorageFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 class BrowsingDataRemoveWebSQLFunction : public BrowsingDataRemoverFunction {
@@ -347,7 +347,7 @@ class BrowsingDataRemoveWebSQLFunction : public BrowsingDataRemoverFunction {
   ~BrowsingDataRemoveWebSQLFunction() override {}
 
   // BrowsingDataRemoverFunction:
-  bool GetRemovalMask(int* removal_mask) override;
+  bool GetRemovalMask(uint64_t* removal_mask) override;
 };
 
 #endif  // CHROME_BROWSER_EXTENSIONS_API_BROWSING_DATA_BROWSING_DATA_API_H_
